% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropHook.R
\name{dropHook}
\alias{dropHook}
\title{drop a function from a hook}
\usage{
dropHook(hookName, n)
}
\arguments{
\item{hookName}{character scalar hook hame; must be one of the names
in \code{names(Motus$hooks)}.  The name indicates the \code{motusClient} function
which is being customized.}

\item{n;}{position of function on hook.  If omitted, all functions are removed
from the hook, otherwise, the `n`th one is removed, if it exists.}
}
\value{
TRUE; if hookName is not valid or there is no hook at position `n` (if supplied),
then stops with an error
}
\description{
drop a function from a hook
}
\note{
: \code{\link{addHook()}} adds functions to the end of a hook, so the
functions on a hook are in order of being added, from earliest to latest.
You can get the list of functions on each hook by examining the global
variable \code{Motus$hooks}.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
