% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{update the metadata for receivers and tags in a motus tag detection database.
only available by tag project ID (not for receiver databases). Updates all existing
metadata previously loaded. This function allows to load all metadata associated 
with a project, instead of only those that are relevant to the detections
in the current database, as per the \link{\code{tagme}} function.}
\usage{
metadata(projID, dir = getwd(), tsBegin = NULL, tsEnd = NULL,
  allRecv = FALSE)
}
\arguments{
\item{projID}{integer scalar project code from motus.org}

\item{tsBegin}{float optional start timestamp to load metadata}

\item{tsEnd}{float optional end timestamp to load metadata}

\item{allRecv}{logical boolean; if TRUE, request metadata for all receivers 
across all projects, based on the user permissions}

\item{dir:}{path to the folder where you are storing databases
Default: the current directory; i.e. \code{getwd()}}
}
\description{
update the metadata for receivers and tags in a motus tag detection database.
only available by tag project ID (not for receiver databases). Updates all existing
metadata previously loaded. This function allows to load all metadata associated 
with a project, instead of only those that are relevant to the detections
in the current database, as per the \link{\code{tagme}} function.
}
\author{
Denis Lepage, Bird Studies Canada
}
