% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvHitsForReceiver.R
\name{srvHitsForReceiver}
\alias{srvHitsForReceiver}
\title{get the hits for a receiver from the data server}
\usage{
srvHitsForReceiver(batchID, hitID = 0)
}
\arguments{
\item{batchID}{integer scalar motus batch ID}

\item{hitID}{double scalar ID of latest hit already obtained.
Default: 0, meaning none.}
}
\value{
data.frame with these columns:
\itemize{
  \item (double) hitID
  \item (double) runID
  \item (integer) batchID
  \item (double) ts
  \item (double) sig
  \item (double) sigSD
  \item (double) noise
  \item (double) freq
  \item (double) freqSD
  \item (double) slop
  \item (double) burstSlop
}
}
\description{
get the hits for a receiver from the data server
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
