% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvRunsForReceiver.R
\name{srvRunsForReceiver}
\alias{srvRunsForReceiver}
\title{get the runs for a receiver from the data server}
\usage{
srvRunsForReceiver(batchID, runID = 0)
}
\arguments{
\item{batchID}{integer scalar motus batch ID}

\item{runID}{double scalar ID of latest run already obtained.
Default: 0, meaning none.}
}
\value{
a data.frame with these columns:
\itemize{
   \item (double) runID
   \item (integer) batchIDbegin
   \item (double) tsBegin
   \item (double) tsEnd
   \item (integer) done
   \item (integer) motusTagID
   \item (integer) ant
   \item (integer) len
}
}
\description{
get the runs for a receiver from the data server
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
