% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyRunsFilter.R
\name{applyRunsFilter}
\alias{applyRunsFilter}
\title{Returns a dataframe of all detections matching the filter conditions. Filters are a list of runID that the user can 
build for the purpose of excluding specific runs from the resulting dataframe.}
\usage{
applyRunsFilter(src, filterName, motusProjID = NA, p.min = 0,
  where.stmt = NA)
}
\arguments{
\item{src}{dplyr sqlite src, as returned by \code{dplyr::src_sqlite()}}

\item{filterName}{unique name given to the filter}

\item{motusProjID}{optional project ID attached to the filter in order to share with other users of the same project.}

\item{p.min}{the minimum probability returned (p.min = 0: no filter applied)}
}
\value{
a dataframe containing the results from alltags, matching the specified filter conditions
}
\description{
Returns a dataframe of all detections matching the filter conditions. Filters are a list of runID that the user can 
build for the purpose of excluding specific runs from the resulting dataframe.
}
\author{
Denis Lepage, Bird Studies Canada
}
