% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyRunsFilter.R
\name{applyRunsFilter}
\alias{applyRunsFilter}
\title{Returns a dataframe of all detections matching the filter conditions. Filters are a list of runID that the user can 
build for the purpose of excluding specific runs from the resulting dataframe.}
\usage{
applyRunsFilter(src, filterNames, motusProjID = NA, p.min = 0,
  where.stmt = NA, p.default = 1)
}
\arguments{
\item{src}{dplyr sqlite src, as returned by \code{dplyr::src_sqlite()}}

\item{motusProjID}{optional project ID attached to the filter in order to share with other users of the same project.}

\item{p.min}{the minimum probability returned (p.min = 0: no filter applied)}

\item{where.stmt}{where statement in SQL format (e.g. where.stmt="motusTagID = 12345 AND runLen >= 4")}

\item{p.default}{default probability (default: 1.0) used and returned by this function for detection records with 
runID's not referenced in the filter. When p.default >= p.min, all records not listed in your filter will be returned.
When p.default < p.min, all records not listed in your filter will be excluded.}

\item{filterName}{unique name given to the filter}
}
\value{
a dplyr tbl object referencing the results from alltags, and including the probability from the specified filter
}
\description{
Returns a dataframe of all detections matching the filter conditions. Filters are a list of runID that the user can 
build for the purpose of excluding specific runs from the resulting dataframe.
}
\author{
Denis Lepage, Bird Studies Canada
}
