% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyRunsFilter.R
\name{applyRunsFilter}
\alias{applyRunsFilter}
\title{Write to the local database the probabilities associated with runs for a filter}
\usage{
applyRunsFilter(src, filterName, motusProjID = NA, p.min = 0,
  where.stmt = NA)
}
\arguments{
\item{src}{dplyr sqlite src, as returned by \code{dplyr::src_sqlite()}}

\item{filterName}{unique name given to the filter}

\item{motusProjID}{optional project ID attached to the filter in order to share with other users of the same project.}

\item{df}{dataframe containing the runID and probability values to save in the local database}

\item{overwrite}{boolean. When TRUE ensures that existing records matching the same filterName 
and runID get replaced}

\item{delete}{boolean. When TRUE, removes all existing filter records associated with the filterName 
and re-inserts the ones contained in the dataframe. This option should be used if the dataframe 
provided contains the entire set of filters you want to save.}
}
\value{
the integer filterID of the filter deleted
}
\description{
Write to the local database the probabilities associated with runs for a filter
}
\author{
Denis Lepage, Bird Studies Canada
}
